Module commitment_dlog::commitment

Re-exports

pub use crate::CommitmentField;

Structs

Challenges

OpeningProof

PolyComm

A polynomial commitment.

Traits

CommitmentCurve

Functions

b_poly

Returns (1 + chal[-1] x)(1 + chal[-2] x^2)(1 + chal[-3] x^4) … It’s “step 8: Define the univariate polynomial” of appendix A.2 of https://eprint.iacr.org/2020/499

b_poly_coefficients

ceil_log2

Returns ceil(log2(d)) but panics if d = 0.

combined_inner_product

Computes the linearization of the evaluations of a (potentially split) polynomial. Each given poly is associated to a matrix where the rows represent the number of evaluated points, and the columns represent potential segments (if a polynomial was split in several parts). Note that if one of the polynomial comes specified with a degree bound, the evaluation for the last segment is potentially shifted to meet the proof.

product

Returns the product of all the field elements belonging to an iterator.

shift_scalar

Inside the circuit, we have a specialized scalar multiplication which computes either