Struct commitment_dlog::srs::SRS

pub struct SRS<G: CommitmentCurve> {
    pub g: Vec<G>,
    pub h: G,
    pub lagrange_bases: HashMap<usize, Vec<G>>,
    pub endo_r: G::ScalarField,
    pub endo_q: G::BaseField,
}

Fields
g: Vec<G>

The vector of group elements for committing to polynomials in coefficient form

h: G

A group element used for blinding commitments

lagrange_bases: HashMap<usize, Vec<G>>

Commitments to Lagrange bases, per domain size

endo_r: G::ScalarField

Coefficient for the curve endomorphism

endo_q: G::BaseField

Coefficient for the curve endomorphism

Implementations

impl<G: CommitmentCurve> SRS<G> where
G::ScalarField: CommitmentField,

pub fn commit(
&self,
plnm: &DensePolynomial<<G as AffineCurve>::ScalarField>,
max: Option<usize>,
rng: &mut impl RngCore + CryptoRng
) -> (PolyComm<G>, PolyComm<<G as AffineCurve>::ScalarField>)

pub fn commit_non_hiding(
&self,
plnm: &DensePolynomial<<G as AffineCurve>::ScalarField>,
max: Option<usize>
) -> PolyComm<G>

pub fn commit_evaluations_non_hiding(
&self,
domain: D<<G as AffineCurve>::ScalarField>,
plnm: &Evaluations<<G as AffineCurve>::ScalarField, D<<G as AffineCurve>::ScalarField>>,
max: Option<usize>
) -> PolyComm<G>

pub fn commit_evaluations(
&self,
domain: D<<G as AffineCurve>::ScalarField>,
plnm: &Evaluations<<G as AffineCurve>::ScalarField, D<<G as AffineCurve>::ScalarField>>,
max: Option<usize>,
rng: &mut impl RngCore + CryptoRng
) -> (PolyComm<G>, PolyComm<<G as AffineCurve>::ScalarField>)

pub fn open<EFqSponge, RNG>(
&self,
group_map: &G::Map,
plnms: Vec<(&DensePolynomial<<G as AffineCurve>::ScalarField>, Option<usize>, PolyComm<<G as AffineCurve>::ScalarField>)>,
elm: &Vec<<G as AffineCurve>::ScalarField>,
polyscale: <G as AffineCurve>::ScalarField,
evalscale: <G as AffineCurve>::ScalarField,
sponge: EFqSponge,
rng: &mut RNG
) -> OpeningProof<G> where
EFqSponge: Clone + FqSponge<<G as AffineCurve>::BaseField, G, <G as AffineCurve>::ScalarField>,
RNG: RngCore + CryptoRng,

pub fn verify<EFqSponge, RNG>(
&self,
group_map: &G::Map,
batch: &mut Vec<(EFqSponge, Vec<<G as AffineCurve>::ScalarField>, <G as AffineCurve>::ScalarField, <G as AffineCurve>::ScalarField, Vec<(&PolyComm<G>, Vec<&Vec<<G as AffineCurve>::ScalarField>>, Option<usize>)>, &OpeningProof<G>)>,
rng: &mut RNG
) -> bool where
EFqSponge: FqSponge<<G as AffineCurve>::BaseField, G, <G as AffineCurve>::ScalarField>,
RNG: RngCore + CryptoRng,

impl<G: CommitmentCurve> SRS<G> where
G::BaseField: PrimeField,
G::ScalarField: CommitmentField,

pub fn max_degree(&self) -> usize

pub fn add_lagrange_basis(&mut self, domain: D<G::ScalarField>)

Compute commitments to the lagrange basis corresponding to the given domain and cache them in the SRS

pub fn create(depth: usize) -> Self

This function creates SRS instance for circuits with number of rows up to depth.

pub fn write<W: Write>(&self, writer: W) -> IoResult<()>

pub fn read<R: Read>(reader: R) -> IoResult<Self>