Struct commitment_dlog::srs::SRS

pub struct SRS<G: CommitmentCurve> {
    pub g: Vec<G>,
    pub h: G,
    pub lagrange_bases: HashMap<usize, Vec<G>>,
    pub endo_r: G::ScalarField,
    pub endo_q: G::BaseField,
}

Fields
g: Vec<G>

The vector of group elements for committing to polynomials in coefficient form

h: G

A group element used for blinding commitments

lagrange_bases: HashMap<usize, Vec<G>>

Commitments to Lagrange bases, per domain size

endo_r: G::ScalarField

Coefficient for the curve endomorphism

endo_q: G::BaseField

Coefficient for the curve endomorphism

Implementations

impl<G: CommitmentCurve> SRS<G>

pub fn commit(
&self,
plnm: &DensePolynomial<<G as AffineCurve>::ScalarField>,
max: Option<usize>,
rng: &mut impl RngCore + CryptoRng
) -> (PolyComm<G>, PolyComm<<G as AffineCurve>::ScalarField>)

Commits a polynomial, potentially splitting the result in multiple commitments.

pub fn mask(
&self,
c: PolyComm<G>,
rng: &mut impl RngCore + CryptoRng
) -> (PolyComm<G>, PolyComm<<G as AffineCurve>::ScalarField>)

Turns a non-hiding polynomial commitment into a hidding polynomial commitment. Transforms each given <a, G> into (<a, G> + wH, w) with a random w per commitment.

pub fn commit_non_hiding(
&self,
plnm: &DensePolynomial<<G as AffineCurve>::ScalarField>,
max: Option<usize>
) -> PolyComm<G>

This function commits a polynomial using the SRS’ basis of size n.

plnm: polynomial to commit to with max size of sections

max: maximal degree of the polynomial (not inclusive), if none, no degree bound The function returns an unbounded commitment vector (which splits the commitment into several commitments of size at most n), as well as an optional bounded commitment (if max is set). Note that a maximum degree cannot (and doesn’t need to) be enforced via a shift if max is a multiple of n.

pub fn commit_helper(
scalars: &[<G as AffineCurve>::ScalarField],
basis: &[G],
n: Option<usize>,
is_zero: bool,
max: Option<usize>
) -> PolyComm<G>

pub fn commit_evaluations_non_hiding(
&self,
domain: D<<G as AffineCurve>::ScalarField>,
plnm: &Evaluations<<G as AffineCurve>::ScalarField, D<<G as AffineCurve>::ScalarField>>,
max: Option<usize>
) -> PolyComm<G>

pub fn commit_evaluations(
&self,
domain: D<<G as AffineCurve>::ScalarField>,
plnm: &Evaluations<<G as AffineCurve>::ScalarField, D<<G as AffineCurve>::ScalarField>>,
max: Option<usize>,
rng: &mut impl RngCore + CryptoRng
) -> (PolyComm<G>, PolyComm<<G as AffineCurve>::ScalarField>)

pub fn verify<EFqSponge, RNG>(
&self,
group_map: &G::Map,
batch: &mut Vec<BatchEvaluationProof<'_, G, EFqSponge>>,
rng: &mut RNG
) -> bool where
EFqSponge: FqSponge<<G as AffineCurve>::BaseField, G, <G as AffineCurve>::ScalarField>,
RNG: RngCore + CryptoRng,
G::BaseField: PrimeField,

This function verifies batch of batched polynomial commitment opening proofs batch: batch of batched polynomial commitment opening proofs vector of evaluation points polynomial scaling factor for this batched openinig proof eval scaling factor for this batched openinig proof batch/vector of polycommitments (opened in this batch), evaluation vectors and, optionally, max degrees opening proof for this batched opening oracle_params: parameters for the random oracle argument randomness source context RETURN: verification status

impl<G: CommitmentCurve> SRS<G>

pub fn open<EFqSponge, RNG>(
&self,
group_map: &G::Map,
plnms: &[(&DensePolynomial<<G as AffineCurve>::ScalarField>, Option<usize>, PolyComm<<G as AffineCurve>::ScalarField>)],
elm: &[<G as AffineCurve>::ScalarField],
polyscale: <G as AffineCurve>::ScalarField,
evalscale: <G as AffineCurve>::ScalarField,
sponge: EFqSponge,
rng: &mut RNG
) -> OpeningProof<G> where
EFqSponge: Clone + FqSponge<<G as AffineCurve>::BaseField, G, <G as AffineCurve>::ScalarField>,
RNG: RngCore + CryptoRng,
G::BaseField: PrimeField,

This function opens polynomial commitments in batch plnms: batch of polynomials to open commitments for with, optionally, max degrees elm: evaluation point vector to open the commitments at polyscale: polynomial scaling factor for opening commitments in batch evalscale: eval scaling factor for opening commitments in batch oracle_params: parameters for the random oracle argument RETURN: commitment opening proof

impl<G: CommitmentCurve> SRS<G> where
G::BaseField: PrimeField,

pub fn max_degree(&self) -> usize

pub fn add_lagrange_basis(&mut self, domain: D<G::ScalarField>)

Compute commitments to the lagrange basis corresponding to the given domain and cache them in the SRS

pub fn create(depth: usize) -> Self

This function creates SRS instance for circuits with number of rows up to depth.